Move the byte set scanner of `extractKeyValuePairs` to `Common/ByteSetLookup.h`
The class is renamed to `ByteSetLookup` and becomes a shared header, so other
byte-class scanners (for example the text index tokenizers) can use it.
`extractKeyValuePairs.cpp` keeps only the extractor itself.
